+++ CHANGE_LOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Better handling of end chat in case of multitab scenarios
 - Prevent new chat creation failure after Proactive chat in Popout mode
 - Fixed post chat having gap in popout mode
+- Fixed logging empty events
 
 ## [1.0.2] - 2023-4-6